moved plugin manager to server/lib/
