+++ server/lib/pluginManager.js
